ScriptPatternTest: move p2shScriptHashFromKeys() from LegacyAddressTest
